refactor(level): split LevelObjects QSC serialization into its own module
level_objects.cpp was 1180 lines. Move QSC serialization (SaveToQSC, ParseTaskLine, UpdateCoordinatesInLine, GenerateTaskLine, SaveSubtreeToQSC, SerializeObjectRecursive, IsSubtreeModified) into level_objects_serialize.cpp. The QSC token formatting/parsing helpers (TaskIdFromArg, EscapeQscString, FormatQscDouble/IntegerToken, ArgTokenFromArg, SplitTopLevelArgs) move to a private level_objects_internal.h shared by both. Core (load/recursive-parse/model-resolution) drops to 519 lines. Also added the new TU to the igi_tests target. Pure code motion: builds clean, runtime-verified (level 3, 1248 objects, no errors).
